backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
Flask-CORS==5.0.0
Flask-Limiter==3.10.0
Flask-WTF==1.2.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==4.1.0
psycopg2-binary==2.9.9
Werkzeug==3.1.0
SQLAlchemy==2.0.36
alembic==1.14.0
PyJWT==2.10.1
bcrypt==4.2.1
WTForms==3.1.2